% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundwater.R
\docType{data}
\name{groundwater}
\alias{groundwater}
\title{Mineral concentrations in ground water}
\format{An object of class \code{character} of length 1.}
\source{
https://twitter.com/beckfrydenborg/status/974787652573646849
}
\usage{
groundwater
}
\description{
Returns the file path to a spreadsheet of data about mineral concentrations
in ground water.  The source was a tweet by Beck Frydenborg about the
difficulty of importing such spreadsheets into R.
}
\keyword{datasets}
