% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\docType{data}
\name{math}
\alias{math}
\title{Math class and school performance}
\format{An object of class \code{character} of length 1.}
\usage{
math
}
\description{
Returns the file path to a spreadsheet of data about the performance of
pupils in a math class and a school.  The source was an education
programme in the USA, which has been anonymised and simplified.
}
\keyword{datasets}
