% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\docType{data}
\name{school}
\alias{school}
\title{School performance}
\format{An object of class \code{character} of length 1.}
\usage{
school
}
\description{
Returns the file path to a spreadsheet of data about the performance of
pupils in a school.  The source was an education programme in the USA, which
has been anonymised and truncated to be manageable examples.
}
\keyword{datasets}
