% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toronto_transit.R
\docType{data}
\name{toronto_transit}
\alias{toronto_transit}
\title{Toronto Transit Commission Ridership Data}
\format{An object of class \code{character} of length 1.}
\source{
Toronto Transit Commission:
https://portal0.cf.opendata.inter.sandbox-toronto.ca/dataset/ttc-ridership-analysis/
}
\usage{
toronto_transit
}
\description{
Returns the file path to a spreadsheet of data about ridership on the Toronto
Transit Commission.
}
\details{
The dataset measures the first point of payment when boarding at the start of
a journey using the TTC and includes ticket types as well as the type of
vehicle.
}
\keyword{datasets}
