% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us-crime.R
\docType{data}
\name{us_crime}
\alias{us_crime}
\alias{us_crime_2}
\alias{us_crime_3}
\title{Crime in the USA}
\format{An object of class \code{character} of length 1.}
\source{
\href{https://ucr.fbi.gov/crime-in-the-u.s/2016/crime-in-the-u.s.-2016/tables/table-2}{Table 2}

\href{https://ucr.fbi.gov/crime-in-the-u.s/2016/crime-in-the-u.s.-2016/tables/table-3}{Table 3}
}
\usage{
us_crime_2

us_crime_3
}
\description{
Return the file paths to two spreadsheets of data about crime in the USA
published by the FBI as part of the Union Crime Reporting programme.
}
\keyword{datasets}
