% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflows.R
\docType{data}
\name{cashflows}
\alias{cashflows}
\title{Fake Cashflow Data}
\format{An object of class \code{character} of length 1.}
\source{
TODO
}
\usage{
cashflows
}
\description{
Returns the file path to a spreadsheet of fake data about cashflows, based on
a real-life spreadsheet.
}
\keyword{datasets}
