% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmarriage.R
\docType{data}
\name{ozmarriage}
\alias{ozmarriage}
\title{Data from the 2017 Australian Marriage Survey}
\format{An object of class \code{character} of length 1.}
\source{
\href{http://www.abs.gov.au/ausstats/abs@.nsf/mf/1800.0}{Australian Bureau ofStatistics}
}
\usage{
ozmarriage
}
\description{
Returns the file path to a spreadsheet of results of the 2017 Australian
Marriage Survey by the Australian Bureau of Statistics, which asked, "Should
the law be changed to allow same-sex couples to marry?"
}
\keyword{datasets}
